rename cmd package to gorillashellout
